# swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DepressoSpeech – Unified API Documentation",
    "description": "Comprehensive API documentation for the DepressoSpeech Depression Severity Prediction System.\n\n## Services\n\n| Service | Port | Description |\n|---------|------|-------------|\n| **MindScope Backend** | 8000 | Core application API – authentication, assessments, file management, doctor/admin dashboards |\n| **DepressoSpeech ML Model** | 8001 | ML inference API – PHQ-8 prediction from audio using eGeMAPS + MFCC + SBERT features |\n| **Frontend (Vite)** | 5173 | React SPA – patient/doctor/admin user interface |\n| **Swagger Docs** | 8080 | This documentation page |\n\n## Authentication\n- **Backend API**: JWT Bearer tokens via `/api/v1/auth/login`\n- **ML Model API**: Optional API key via `X-API-Key` header\n- **Admin**: Separate admin login via `/api/v1/auth/admin/login`\n\n## Architecture\nThe system follows a microservices pattern:\n1. Frontend sends requests to Backend (proxied via Vite in dev)\n2. Backend orchestrates business logic, stores data in SQLite\n3. Backend delegates ML inference to the Model service\n4. Model service processes audio and returns PHQ-8 predictions",
    "version": "2.0.0",
    "contact": {
      "name": "DepressoSpeech Team – KMIT",
      "email": "koushik@kmit.in"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
...
    },
    {
...
    }
  ],
...
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register a new user",
        "description": "Register a new patient or doctor account. An OTP is sent to the provided email for verification.",
        "operationId": "register",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" },
              "examples": {
                "patient": {
                  "summary": "Register as patient",
                  "value": {
                    "role": "patient",
                    "name": "John Doe",
                    "email": "john@example.com",
                    "password": "securePass123",
                    "age": 28,
                    "basicInfo": "Student, first-time screening"
                  }
                },
                "doctor": {
                  "summary": "Register as doctor",
                  "value": {
                    "role": "doctor",
                    "name": "Dr. Smith",
                    "email": "smith@hospital.com",
                    "password": "docSecure456",
                    "specialization": "Psychiatry",
                    "licenseNumber": "PSY-12345",
                    "clinicName": "City Mental Health",
                    "yearsExperience": 10
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created successfully. OTP sent for email verification.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RegisterResponse" }
              }
            }
          },
          "409": { "description": "Email already registered" },
          "422": { "description": "Validation error" }
        }
      }
    },
    "/auth/verify-otp": {
      "post": {
        "tags": ["Auth"],
        "summary": "Verify email OTP",
        "description": "Verify user email with the 6-digit OTP sent during registration.",
        "operationId": "verifyOtp",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/VerifyOtpRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Email verified successfully" },
          "400": { "description": "Invalid OTP" },
          "404": { "description": "Account not found" },
          "410": { "description": "OTP expired" }
        }
      }
    },
    "/auth/resend-otp": {
      "post": {
        "tags": ["Auth"],
        "summary": "Resend OTP",
        "description": "Generate and send a new OTP to the user's email.",
        "operationId": "resendOtp",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ResendOtpRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "New OTP sent" },
          "404": { "description": "Account not found" },
          "500": { "description": "Failed to send email" }
        }
      }
    },
    "/auth/forgot-password": {
      "post": {
        "tags": ["Auth"],
        "summary": "Request password reset",
        "description": "Send an OTP to the user's email for password reset. Does not reveal whether the email exists.",
        "operationId": "forgotPassword",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ForgotPasswordRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "OTP sent if email is registered" }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "tags": ["Auth"],
        "summary": "Reset password with OTP",
        "description": "Reset user password after validating the OTP received via email.",
        "operationId": "resetPassword",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ResetPasswordRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Password reset successfully" },
          "400": { "description": "Invalid OTP" },
          "404": { "description": "Account not found" },
          "410": { "description": "OTP expired" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login with email and password",
        "description": "Authenticate with email and password. Returns JWT access and refresh tokens.",
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" }
              }
            }
          },
          "401": { "description": "Invalid credentials" },
          "423": { "description": "Account locked" }
        }
      }
    },
    "/auth/admin/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Admin login",
        "description": "Authenticate as system administrator.",
        "operationId": "adminLogin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AdminLoginRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Admin login successful" },
          "401": { "description": "Invalid admin credentials" }
        }
      }
    },
    "/auth/google": {
      "post": {
        "tags": ["Auth"],
        "summary": "Google OAuth login",
        "description": "Authenticate with a Google ID token from Google Sign-In.",
        "operationId": "googleLogin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/GoogleAuthRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Google login successful" },
          "401": { "description": "Invalid Google token" },
          "501": { "description": "Google OAuth not configured" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Logout",
        "description": "Invalidate user session.",
        "operationId": "logout",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LogoutRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Logged out successfully" }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Get current user",
        "description": "Retrieve the currently authenticated user's profile.",
        "operationId": "getCurrentUser",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Current user profile",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          },
          "401": { "description": "Not authenticated" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from `/api/v1/auth/login`"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Optional API key for ML model (set via DEPRESSO_API_KEY env var)"
      }
    },
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "required": ["role", "name", "email", "password"],
        "properties": {
          "role": { "type": "string", "enum": ["patient", "doctor"], "description": "User role" },
          "name": { "type": "string", "minLength": 1, "maxLength": 120 },
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 6 },
          "age": { "type": "integer", "description": "Patient only" },
          "basicInfo": { "type": "string", "description": "Patient only" },
          "specialization": { "type": "string", "description": "Doctor only" },
          "licenseNumber": { "type": "string", "description": "Doctor only" },
          "clinicName": { "type": "string", "description": "Doctor only" },
          "yearsExperience": { "type": "integer", "description": "Doctor only" }
        }
      },
      "RegisterResponse": {
        "type": "object",
        "properties": {
          "user": { "$ref": "#/components/schemas/UserObject" },
          "requiresVerification": { "type": "boolean", "example": true },
          "message": { "type": "string" }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "accessToken": { "type": "string" },
          "refreshToken": { "type": "string" },
          "expiresIn": { "type": "integer", "description": "Token TTL in seconds" },
          "user": { "$ref": "#/components/schemas/UserObject" }
        }
      },
      "AdminLoginRequest": {
        "type": "object",
        "required": ["adminId", "password"],
        "properties": {
          "adminId": { "type": "string" },
          "password": { "type": "string" }
        }
      },
      "GoogleAuthRequest": {
        "type": "object",
        "required": ["credential"],
        "properties": {
          "credential": { "type": "string", "minLength": 1, "description": "Google ID token" }
        }
      },
      "VerifyOtpRequest": {
        "type": "object",
        "required": ["email", "otp"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "otp": { "type": "string", "minLength": 6, "maxLength": 6 }
        }
      },
      "ResendOtpRequest": {
        "type": "object",
        "required": ["email"],
        "properties": {
          "email": { "type": "string", "format": "email" }
        }
      },
      "ForgotPasswordRequest": {
        "type": "object",
        "required": ["email"],
        "properties": {
          "email": { "type": "string", "format": "email" }
        }
      },
      "ResetPasswordRequest": {
        "type": "object",
        "required": ["email", "otp", "newPassword"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "otp": { "type": "string", "minLength": 6, "maxLength": 6 },
          "newPassword": { "type": "string", "minLength": 6 }
        }
      },
      "LogoutRequest": {
        "type": "object",
        "properties": {
          "refreshToken": { "type": "string" }
        }
      },
      "UserObject": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "role": { "type": "string", "enum": ["patient", "doctor", "admin"] },
          "name": { "type": "string" },
          "email": { "type": "string" },
          "isVerified": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "age": { "type": "integer" },
          "basicInfo": { "type": "string" },
          "specialization": { "type": "string" },
          "licenseNumber": { "type": "string" },
          "clinicName": { "type": "string" },
          "yearsExperience": { "type": "integer" }
        }
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "user": { "$ref": "#/components/schemas/UserObject" }
        }
      },
...
    }
  }
}